refactor: add UI directories and restructure architecture
