+++ src/index.html
@@ -7,7 +7,7 @@
   </head>
   <body>
     <div id="app">
-      <script type="module" src="./app/route.ts"></script>
+      <script type="module" src="./app/routes/route.ts"></script>
     </div>
   </body>
 </html>
